Split request handling code into different classes
