+++ tests/tempo.test.ts
@@ -19,25 +19,6 @@ function getOptions(options?: any) {
 }
 
 describe('TempoAi', () => {
-  describe('Constructor Tests', () => {
-    it('Constructor functions properly', () => {
-      const tempo = new TempoApi(
-        getOptions({
-          timeout: 3000
-        })
-      );
-
-      expect(tempo.protocol).toEqual('http');
-      expect(tempo.host).toEqual('tempo.somehost.com');
-      expect(tempo.port).toEqual('8080');
-      expect(tempo.baseOptions.auth.user).toEqual('');
-      expect(tempo.baseOptions.auth.pass).toEqual('');
-      expect(tempo.baseOptions.auth.bearer).toEqual('sometoken');
-      expect(tempo.baseOptions.timeout).toEqual(3000);
-      expect(tempo.apiVersion).toEqual('3');
-    });
-  });
-
   describe('Request Functions Tests', () => {
     async function dummyURLCall(
       tempoApiMethodName: string,
